requirements-tests.txt
coverage==4.5.4
coveralls==1.8.2
selenium==3.141.0
locustio==0.13.3
